Allow base64 decode ignore full stop and fix non-mime text bug
Base64 mime mail body should not contain the full stop (.) char
as that is not a valid base64 char. Now this fullstop is ignored
during base64 decoding.

There was a bug where a non-mime mail body contains only plain
text body that requires conversion to utf-8 is not converted to
utf-8. This commit fixes the bug and now this type of mail body
text will be converted to utf-8 text properly.

Add base64 content with full stop decode test. Also add parse test
for non-mime text body that is not utf-8 and full of russian text.